refactor: move useStyling hook
